Move the root-level Jobs-app test classes into Apps/Jobs (00439)
JobsAppRefreshGatingTests and JobsAppPromptDisplayTests sat at the test project root while JobsTableColumnLabelTests and JobsTableFilterTests already lived in the mirrored Apps/Jobs folder. Move both and reparent their namespace to Ivy.Tendril.Test.Apps.Jobs; the project globs *.cs and neither class is referenced elsewhere, so nothing else changes.
