+++ src/Ivy.Tendril.Test/Apps/Jobs/JobsAppPromptDisplayTests.cs
@@ -1,7 +1,7 @@
 using Ivy.Tendril.Apps.Jobs;
 using Ivy.Tendril.Models;
 
-namespace Ivy.Tendril.Test;
+namespace Ivy.Tendril.Test.Apps.Jobs;
 
 public class JobsAppPromptDisplayTests
 {
